+++ examples/advanced/src/gen/msw/petController/findPetsByStatusHandler.ts
@@ -1,5 +1,5 @@
-import { http } from 'msw'
 import type { FindPetsByStatusQueryResponse } from '../../models/ts/petController/FindPetsByStatus.ts'
+import { http } from 'msw'
 
 export function findPetsByStatusHandler(data?: FindPetsByStatusQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response)) {
   return http.get('*/pet/findByStatus/:step_id', function handler(info) {