+++ packages/plugin-svelte-query/src/components/Mutation.tsx
@@ -19,21 +19,23 @@ type Props = {
   mutationKeyName: string
   typeSchemas: OperationSchemas
   operation: Operation
+  paramsCasing: PluginSvelteQuery['resolvedOptions']['paramsCasing']
   paramsType: PluginSvelteQuery['resolvedOptions']['paramsType']
   dataReturnType: PluginSvelteQuery['resolvedOptions']['client']['dataReturnType']
   pathParamsType: PluginSvelteQuery['resolvedOptions']['pathParamsType']
 }
 
 type GetParamsProps = {
+  paramsCasing: PluginSvelteQuery['resolvedOptions']['paramsCasing']
   pathParamsType: PluginSvelteQuery['resolvedOptions']['pathParamsType']
   dataReturnType: PluginSvelteQuery['resolvedOptions']['client']['dataReturnType']
   typeSchemas: OperationSchemas
 }
 
-function getParams({ dataReturnType, typeSchemas }: GetParamsProps) {
+function getParams({ paramsCasing, dataReturnType, typeSchemas }: GetParamsProps) {
   const TData = dataReturnType === 'data' ? typeSchemas.response.name : `ResponseConfig<${typeSchemas.response.name}>`
   const mutationParams = FunctionParams.factory({
-    ...getPathParams(typeSchemas.pathParams, { typed: true }),
+    ...getPathParams(typeSchemas.pathParams, { typed: true, casing: paramsCasing }),
     data: typeSchemas.request?.name
       ? {
           type: typeSchemas.request?.name,
@@ -68,25 +70,37 @@ function getParams({ dataReturnType, typeSchemas }: GetParamsProps) {
   })
 }
 
-export function Mutation({ name, clientName, paramsType, pathParamsType, dataReturnType, typeSchemas, operation, mutationKeyName }: Props): ReactNode {
+export function Mutation({
+  name,
+  clientName,
+  paramsCasing,
+  paramsType,
+  pathParamsType,
+  dataReturnType,
+  typeSchemas,
+  operation,
+  mutationKeyName,
+}: Props): ReactNode {
   const mutationKeyParams = MutationKey.getParams({
     pathParamsType,
     typeSchemas,
   })
 
   const params = getParams({
+    paramsCasing,
     pathParamsType,
     dataReturnType,
     typeSchemas,
   })
 
   const clientParams = Client.getParams({
+    paramsCasing,
     paramsType,
     typeSchemas,
     pathParamsType,
   })
   const mutationParams = FunctionParams.factory({
-    ...getPathParams(typeSchemas.pathParams, { typed: true }),
+    ...getPathParams(typeSchemas.pathParams, { typed: true, casing: paramsCasing }),
     data: typeSchemas.request?.name
       ? {
           type: typeSchemas.request?.name,