+++ packages/plugin-svelte-query/src/components/Query.tsx
@@ -18,19 +18,21 @@ type Props = {
   queryKeyTypeName: string
   typeSchemas: OperationSchemas
   operation: Operation
+  paramsCasing: PluginSvelteQuery['resolvedOptions']['paramsCasing']
   paramsType: PluginSvelteQuery['resolvedOptions']['paramsType']
   pathParamsType: PluginSvelteQuery['resolvedOptions']['pathParamsType']
   dataReturnType: PluginSvelteQuery['resolvedOptions']['client']['dataReturnType']
 }
 
 type GetParamsProps = {
+  paramsCasing: PluginSvelteQuery['resolvedOptions']['paramsCasing']
   paramsType: PluginSvelteQuery['resolvedOptions']['paramsType']
   pathParamsType: PluginSvelteQuery['resolvedOptions']['pathParamsType']
   dataReturnType: PluginSvelteQuery['resolvedOptions']['client']['dataReturnType']
   typeSchemas: OperationSchemas
 }
 
-function getParams({ paramsType, pathParamsType, dataReturnType, typeSchemas }: GetParamsProps) {
+function getParams({ paramsType, paramsCasing, pathParamsType, dataReturnType, typeSchemas }: GetParamsProps) {
   const TData = dataReturnType === 'data' ? typeSchemas.response.name : `ResponseConfig<${typeSchemas.response.name}>`
 
   if (paramsType === 'object') {
@@ -38,7 +40,7 @@ function getParams({ paramsType, pathParamsType, dataReturnType, typeSchemas }: 
       data: {
         mode: 'object',
         children: {
-          ...getPathParams(typeSchemas.pathParams, { typed: true }),
+          ...getPathParams(typeSchemas.pathParams, { typed: true, casing: paramsCasing }),
           data: typeSchemas.request?.name
             ? {
                 type: typeSchemas.request?.name,
@@ -75,7 +77,7 @@ function getParams({ paramsType, pathParamsType, dataReturnType, typeSchemas }: 
     pathParams: typeSchemas.pathParams?.name
       ? {
           mode: pathParamsType === 'object' ? 'object' : 'inlineSpread',
-          children: getPathParams(typeSchemas.pathParams, { typed: true }),
+          children: getPathParams(typeSchemas.pathParams, { typed: true, casing: paramsCasing }),
           optional: isOptional(typeSchemas.pathParams?.schema),
         }
       : undefined,
@@ -115,6 +117,7 @@ export function Query({
   queryOptionsName,
   queryKeyName,
   paramsType,
+  paramsCasing,
   pathParamsType,
   dataReturnType,
   typeSchemas,
@@ -126,15 +129,18 @@ export function Query({
 
   const queryKeyParams = QueryKey.getParams({
     pathParamsType,
+    paramsCasing,
     typeSchemas,
   })
   const queryOptionsParams = QueryOptions.getParams({
     paramsType,
+    paramsCasing,
     pathParamsType,
     typeSchemas,
   })
   const params = getParams({
     paramsType,
+    paramsCasing,
     pathParamsType,
     dataReturnType,
     typeSchemas,