+++ packages/plugin-svelte-query/src/generators/mutationGenerator.test.tsx
@@ -85,6 +85,7 @@ describe('mutationGenerator operation', async () => {
         importPath: '@kubb/plugin-client/client',
       },
       parser: 'zod',
+      paramsCasing: undefined,
       paramsType: 'inline',
       pathParamsType: 'inline',
       queryKey: QueryKey.getTransformer,