+++ packages/plugin-swr/src/generators/queryGenerator.test.tsx
@@ -98,6 +98,7 @@ describe('queryGenerator operation', async () => {
         dataReturnType: 'data',
         importPath: '@kubb/plugin-client/client',
       },
+      paramsCasing: undefined,
       queryKey: QueryKey.getTransformer,
       mutationKey: MutationKey.getTransformer,
       query: {