+++ packages/plugin-swr/src/types.ts
@@ -7,6 +7,7 @@ import type { PluginReactQuery } from '@kubb/plugin-react-query'
 type TransformerProps = {
   operation: Operation
   schemas: OperationSchemas
+  casing: 'camelcase' | undefined
 }
 
 export type Transformer = (props: TransformerProps) => unknown[]
@@ -81,6 +82,11 @@ export type Options = {
   mutationKey?: MutationKey
   mutation?: Mutation | false
   /**
+   * How to style your params, by default no casing is applied
+   * - 'camelcase' will use camelcase for the params names
+   */
+  paramsCasing?: 'camelcase'
+  /**
    * How to pass your params
    * - 'object' will return the params and pathParams as an object.
    * - 'inline' will return the params as comma separated params.
@@ -119,6 +125,7 @@ type ResolvedOptions = {
   query: NonNullable<Required<Query>> | false
   mutationKey: MutationKey | undefined
   mutation: NonNullable<Required<Mutation>> | false
+  paramsCasing: Options['paramsCasing']
   paramsType: NonNullable<Options['paramsType']>
   pathParamsType: NonNullable<Options['pathParamsType']>
 }